Split test helpers to mocks.h
